Add files via upload
Here is a deluge function to generate a random string, such as a password or encryption key programmatically. This snippet allows you to set the password length. You can also add or remove options from the “specials” array to customize the special characters that could be used. 

One thing to note is that this snippet does not guarantee that there will be a certain number of special characters or letters or numbers. You could enhance this script yourself by adding regex matches() checks to test if your generated string matches your requirements.

https://club.zenatta.com/c/code-share/random-string-generator